:tada: fix some stuff

+++ src/events/client/ready.js
@@ -28,36 +28,38 @@ module.exports = {
     const members = client.utils.formatNumber(
       client.guilds.cache.reduce((a, b) => a + b.memberCount, 0)
     );
-    const commands = client.commands.size;
-    const boot = client.bootTime;
-    const message = `${icon} \`[ ${client.version} ]\` **REBOOT**`;
-    const embed = {
-      color: "GREY",
-      description: [
-        "```properties",
-        `Servers: ${servers}`,
-        `Members: ${members}`,
-        `Command: ${commands}`,
-        `Boot: ${boot}ms`,
-        "```",
-      ].join("\n"),
-    };
+    require("../../helpers/pingChecker")(client);
+    if (!client.config.dev) {
+      const commands = client.commands.size;
+      const boot = client.bootTime;
+      const message = `${icon} \`[ ${client.version} ]\` **REBOOT**`;
+      const embed = {
+        color: "GREY",
+        description: [
+          "```properties",
+          `Servers: ${servers}`,
+          `Members: ${members}`,
+          `Command: ${commands}`,
+          `Boot: ${boot}ms`,
+          "```",
+        ].join("\n"),
+      };
 
-    await client.channels.cache
-      .get("")
-      ?.createWebhook(bot, {
-        avatar: client.user.displayAvatarURL({
-          format: "png",
-          dynamic: true,
-          size: 128,
-        }),
-      })
-      .then((webhook) =>
-        Promise.all([webhook.send(message, { embeds: [embed] }), webhook])
-      )
-      .then(([_, webhook]) => webhook.delete())
-      .catch(() => {});
-
+      await client.channels.cache
+        .get("803706448784785488")
+        ?.createWebhook(bot, {
+          avatar: client.user.displayAvatarURL({
+            format: "png",
+            dynamic: true,
+            size: 128,
+          }),
+        })
+        .then((webhook) =>
+          Promise.all([webhook.send(message, { embeds: [embed] }), webhook])
+        )
+        .then(([_, webhook]) => webhook.delete())
+        .catch(() => {});
+    }
     for (const guild of client.guilds.cache) {
       setInterval(async () => {
         await autoCovid(client, guild.id);
